https://github.com/ledwindra/replication-code-economics/issues/4

+++ src/aea_metadata.py
@@ -3,6 +3,7 @@ import os
 import requests
 from bs4 import BeautifulSoup
 from datetime import datetime
+from time import sleep
 
 def current_timestamp():
     '''
@@ -29,42 +30,54 @@ def get_metadata(url):
     Download ICPSR metadata (Dublin Core)
     '''
 
-    res = requests.get(url)
-    content = BeautifulSoup(res.content, features='html.parser')
-    well = content.find_all('div', {'class': 'well'})
-    for i in range(len(well)):
+    status_code = []
+    while status_code != 200:
         try:
-            if well[i].find('h2', {'class': 'asideH2'}).text == 'Export Metadata':
-                break
-        except AttributeError:
-            pass
-
-    well = well[i]
-    metadata = [x['href'] for x in well.find_all('a')]
-
-    return metadata[0]
+            res = requests.get(url, timeout=5)
+            status_code = res.status_code
+            content = BeautifulSoup(res.content, features='html.parser')
+            well = content.find_all('div', {'class': 'well'})
+            for i in range(len(well)):
+                try:
+                    if well[i].find('h2', {'class': 'asideH2'}).text == 'Export Metadata':
+                        well = well[i]
+                        metadata = [x['href'] for x in well.find_all('a')]
+                        return metadata[0]
+                except AttributeError:
+                    pass
+        except Exception as e:
+            print(f'{current_timestamp()}: {e}')
+            sleep(3)
 
 def parse_metadata(metadata):
     '''
     Parses metadata and returns into a JSON file
     '''
 
-    res = requests.get(metadata)
-    content = BeautifulSoup(res.content, features='xml')
-    try:
-        file_name = content.find('identifier').text
-        metadata = {
-            'datestamp': content.find('datestamp').text,
-            'title': content.find('title').text,
-            'creator': content.find('creator').text,
-            'identifier': [x.text for x in content.find_all('identifier')],
-            'description': content.find('description').text,
-            'subject': [x.text for x in content.find_all('subject')]
-        }
-        with open(f'data/aea-metadata/{file_name}.json', 'w') as f:
-            json.dump(metadata, f, indent=4)
-    except AttributeError:
-        pass
+    status_code = []
+    while status_code != 200:
+        try:
+            res = requests.get(metadata, timeout=5)
+            status_code = res.status_code
+            content = BeautifulSoup(res.content, features='xml')
+            try:
+                file_name = content.find('identifier').text
+                if not os.path.exists(f'data/aea-metadata/{file_name}.json'):
+                    metadata = {
+                        'datestamp': content.find('datestamp').text,
+                        'title': content.find('title').text,
+                        'creator': content.find('creator').text,
+                        'identifier': [x.text for x in content.find_all('identifier')],
+                        'description': content.find('description').text,
+                        'subject': [x.text for x in content.find_all('subject')]
+                    }
+                    with open(f'data/aea-metadata/{file_name}.json', 'w') as f:
+                        json.dump(metadata, f, indent=4)
+            except AttributeError:
+                pass
+        except Exception as e:
+            print(f'{current_timestamp()}: {e}')
+            sleep(3)
 
 if __name__ == '__main__':
     aea = os.listdir('data/aea/')
